Revert "Add Other Logging Implementations (#858)"
This reverts commit b7b31c45b0eb350deedd282b88398d1ca0d5faf4.
