+++ launch/launch/actions/__init__.py
@@ -23,11 +23,7 @@ from .for_loop import ForEach
 from .for_loop import ForLoop
 from .group_action import GroupAction
 from .include_launch_description import IncludeLaunchDescription
-from .log import Log
-from .log import LogDebug
-from .log import LogError
-from .log import LogInfo
-from .log import LogWarning
+from .log_info import LogInfo
 from .opaque_coroutine import OpaqueCoroutine
 from .opaque_function import OpaqueFunction
 from .pop_environment import PopEnvironment
@@ -55,11 +51,7 @@ __all__ = [
     'ForLoop',
     'GroupAction',
     'IncludeLaunchDescription',
-    'Log',
-    'LogDebug',
-    'LogError',
     'LogInfo',
-    'LogWarning',
     'OpaqueCoroutine',
     'OpaqueFunction',
     'PopEnvironment',
